Extend frame V2 data and cut down fields
Still WIP,

But added some more outputs of some Frame V2 data, for some of the "Control transfer" mode of operations

I also updated the names of many of the fields to use generic names "value" and "value2"

This made the data table much smaller and as such easier to see all of the data.

I also updated my HLA using this to the new names.

I still have more work to do in the Control transfer mode, but will get there when I get there.